added onboarding
An onboarding screen is a type of user interface (UI) element that is presented to a user when they first start using a new software application or service. It is designed to help users quickly get up to speed with the features and functionality of the application, and may include tutorials, tips, and other interactive elements to help users learn how to use the app effectively. Onboarding screens are often used to introduce users to the core features of an app, and may also provide information about account setup, privacy and security, and other important details.