chore: Ignore firebase_config.dart for enhanced project security
By adding firebase_config.dart to the .gitignore file, we are ensuring that the file containing sensitive API keys and configuration details for Firebase is not tracked by Git and will not be included in future commits. This step enhances the security of our project by preventing the exposure of sensitive information on public repositories. Developers should follow the provided instructions to set up their own local firebase_config.dart file with the necessary credentials.