feat: Add initial instructions and firebase_config.dart template
This commit adds the initial instructions for getting started with the Wiser project. It provides step-by-step guidelines for cloning the repository, generating the `firebase_config.dart` file using the Firebase CLI, creating the necessary file structure, and pasting the code snippet into the `firebase_config.dart` file.

A note is included to highlight the importance of not including sensitive information in the repository and creating a private `firebase_config.dart` file with actual Firebase project details.

Additionally, a cute GIF image is included to add some visual appeal to the documentation.